Made changes in ofxSmile
Added mousePressed function
Attempting to make it so when you finish a video or watched an image
for 5 seconds or webpage for 30 seconds that you the program
automatically starts looking for smiles. Once it has found a smile it
returns a true boolean. Based on this boolean the main program should
start running again. In the end it might be more user-friendly to  just
use the mouse.